pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-news-agent"
version = "0.1.0"
description = "LangGraph-based AI news digest agent for Telegram"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "langgraph",
  "langchain",
  "langsmith",
  "langgraphics @ git+https://github.com/proactive-agent/langgraphics.git@main",
  "httpx",
  "feedparser",
  "beautifulsoup4",
  "lxml",
  "pydantic>=2,<3",
  "pydantic-settings>=2,<3",
  "python-dotenv",
  "PyYAML",
  "python-dateutil",
]

[project.optional-dependencies]
dev = [
  "pytest",
  "pytest-asyncio",
  "mypy",
  "ruff",
]

[project.scripts]
ai-news-agent = "app.main:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
no_implicit_optional = true
